authenticated user to delete post

+++ Controller/api/v1/post_api.js
@@ -33,14 +33,26 @@ module.exports.postlist = async function(req,res){
 module.exports.destory = async function(req,res){
     try{
         let post = await Post.findById(req.params.id);
-        post.remove();
-        await Comment.deleteMany({post : req.params.id});
-        return res.json(200,{
-            message:"deleted post successfully"
-        })
+        if(post.user == req.user.id){
+            post.remove();
+            await Comment.deleteMany({post : req.params.id});
+            return res.json(200,{
+                message:"deleted post successfully"
+            })
+        }
+        
+        else{
+            console.log(post,req.user.id);
+            return res.json(404,{
+                message:"unauthorized"
+            })
+        }
     }
     catch(err){
-
+        console.log(err);
+        return res.json(404,{
+            message:"Not authorized to delete post"
+        })
     }
 }
 module.exports.createpost = async function(req,res){
